backend/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
fastapi==0.115.2
uvicorn==0.30.6
